Replace double with long long for cumulative
git-svn-id: file:///Users/tack/GecodeGitMigration/gecode-svn-mirror/gecode/trunk@13487 e85b7adc-8362-4630-8c63-7469d557c915

+++ gecode/int/cumulative/overload.hpp
@@ -52,7 +52,7 @@ namespace Gecode { namespace Int { namespace Cumulative {
     for (int i=0; i<f.size(); i++) {
       if (f[i].mandatory())
         o.insert(i);
-      if (o.env() > static_cast<double>(c)*f[i].lct())
+      if (o.env() > static_cast<long long int>(c)*f[i].lct())
         GECODE_ME_CHECK(f[i].excluded(home));
     }
     return ES_OK;
